backend/flask_model/requirements.txt
flask==3.0.3
flask-cors==4.0.1
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
Pillow==10.3.0
opencv-python==4.9.0.80
